added healthcheck listening (doesnt work ffs)

+++ server/internal/entities/platform.go
@@ -2,16 +2,16 @@ package entities
 
 type Platform struct {
 	coord    Vector
-	width    int
-	height   int
+	width    float64
+	height   float64
 	velocity float64
 }
 
-func (p Platform) Height() int {
+func (p Platform) Height() float64 {
 	return p.height
 }
 
-func (p Platform) Width() int {
+func (p Platform) Width() float64 {
 	return p.width
 }
 
@@ -27,7 +27,7 @@ func (p *Platform) Move(offset float64) {
 	p.coord.Y += offset
 }
 
-func NewPlatform(width int, height int, coord Vector) *Platform {
+func NewPlatform(width float64, height float64, coord Vector) *Platform {
 	p := new(Platform)
 	p.width = width
 	p.height = height
